move code to util module
