+++ gstd/src/lib.rs
@@ -131,6 +131,8 @@
 #![cfg_attr(feature = "strict", deny(warnings))]
 #![doc(html_logo_url = "https://docs.gear.rs/logo.svg")]
 #![doc(test(attr(deny(warnings), allow(unused_variables, unused_assignments))))]
+#![allow(incomplete_features)]
+#![feature(specialization)]
 
 extern crate alloc;
 #[cfg(target_arch = "wasm32")]
